requirements.txt
streamlit
langchain_community
langchain_text_splitters
langchain_groq
langchain_openai
langchain_core
langchain
validators
python-dotenv
faiss-cpu